requirements.txt
aspose-words==22.10.0